requirements.txt
gradio
openai-whisper
ffmpeg-python
numpy==1.23
